Added new page for adding new restaurants exclusive of the Django Admin Add Restaurant Form is created via the Restaurant Model
Restaurant Address is Geocoded through the (free) Google Geocoding API
via a string of Address+City+State+Country (State is send as blank when
a non-US or Canadian country is used).

Google Geocoding has 4 levels of accuracy. Anything but the top level of
accuracy returns an error and does not save the entry.